@@ -1,11 +1,11 @@
 name: image_crop
 description: A flutter plugin to crop image on iOS and Android. It processes image files off main thread natively. The plugin provides a Crop widget to display image cropping to a user.
-version: 0.3.4
+version: 0.4.0
 homepage: https://github.com/VolodymyrLykhonis/image_crop
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.6 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.6"
 
 dependencies:
   flutter:
