@@ -15,15 +15,15 @@ enum _CropHandleSide { none, topLeft, topRight, bottomLeft, bottomRight }
 
 class Crop extends StatefulWidget {
   final ImageProvider image;
-  final double aspectRatio;
+  final double? aspectRatio;
   final double maximumScale;
   final bool alwaysShowGrid;
   final ImageErrorListener onImageError;
   final Function onImageSuccess;
 
   const Crop({
-    Key key,
-    this.image,
+    Key? key,
+    required this.image,
     this.aspectRatio,
     this.maximumScale = 2.0,
     this.alwaysShowGrid = false,
@@ -36,7 +36,7 @@ class Crop extends StatefulWidget {
 
   Crop.file(
     File file, {
-    Key key,
+    Key? key,
     double scale = 1.0,
     this.aspectRatio,
     this.maximumScale = 2.0,
@@ -44,66 +44,63 @@ class Crop extends StatefulWidget {
     this.onImageError,
     this.onImageSuccess,
   })  : image = FileImage(file, scale: scale),
-        assert(maximumScale != null),
-        assert(alwaysShowGrid != null),
         super(key: key);
 
   Crop.asset(
     String assetName, {
-    Key key,
-    AssetBundle bundle,
-    String package,
+    Key? key,
+    AssetBundle? bundle,
+    String? package,
     this.aspectRatio,
     this.maximumScale = 2.0,
     this.alwaysShowGrid = false,
     this.onImageError,
     this.onImageSuccess,
   })  : image = AssetImage(assetName, bundle: bundle, package: package),
-        assert(maximumScale != null),
-        assert(alwaysShowGrid != null),
         super(key: key);
 
   @override
   State<StatefulWidget> createState() => CropState();
 
-  static CropState of(BuildContext context) {
-    return context.findAncestorStateOfType<CropState>();
-  }
+  static CropState? of(BuildContext context) =>
+      context.findAncestorStateOfType<CropState>();
 }
 
 class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
   final _surfaceKey = GlobalKey();
-  AnimationController _activeController;
-  AnimationController _settleController;
-  ImageStream _imageStream;
-  ui.Image _image;
-  double _scale;
-  double _ratio;
-  Rect _view;
-  Rect _area;
-  Offset _lastFocalPoint;
-  _CropAction _action;
-  _CropHandleSide _handle;
-  double _startScale;
-  Rect _startView;
-  Tween<Rect> _viewTween;
-  Tween<double> _scaleTween;
-  ImageStreamListener _imageListener;
+
+  late final AnimationController _activeController;
+  late final AnimationController _settleController;
+
+  double _scale = 1.0;
+  double _ratio = 1.0;
+  Rect _view = Rect.zero;
+  Rect _area = Rect.zero;
+  Offset _lastFocalPoint = Offset.zero;
+  _CropAction _action = _CropAction.none;
+  _CropHandleSide _handle = _CropHandleSide.none;
+
+  late double _startScale;
+  late Rect _startView;
+  late Tween<Rect?> _viewTween;
+  late Tween<double> _scaleTween;
+
+  ImageStream? _imageStream;
+  ui.Image? _image;
+  ImageStreamListener? _imageListener;
 
   double get scale => _area.shortestSide / _scale;
 
-  Rect get area {
-    return _view.isEmpty
-        ? null
-        : Rect.fromLTWH(
-            _area.left * _view.width / _scale - _view.left,
-            _area.top * _view.height / _scale - _view.top,
-            _area.width * _view.width / _scale,
-            _area.height * _view.height / _scale,
-          );
-  }
+  Rect? get area => _view.isEmpty
+      ? null
+      : Rect.fromLTWH(
+          _area.left * _view.width / _scale - _view.left,
+          _area.top * _view.height / _scale - _view.top,
+          _area.width * _view.width / _scale,
+          _area.height * _view.height / _scale,
+        );
 
-  bool get _isEnabled => !_view.isEmpty && _image != null;
+  bool get _isEnabled => _view.isEmpty == false && _image != null;
 
   // Saving the length for the widest area for different aspectRatio's
   final Map<double, double> _maxAreaWidthMap = {};
@@ -114,13 +111,7 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
   @override
   void initState() {
     super.initState();
-    _area = Rect.zero;
-    _view = Rect.zero;
-    _scale = 1.0;
-    _ratio = 1.0;
-    _lastFocalPoint = Offset.zero;
-    _action = _CropAction.none;
-    _handle = _CropHandleSide.none;
+
     _activeController = AnimationController(
       vsync: this,
       value: widget.alwaysShowGrid ? 1.0 : 0.0,
@@ -131,21 +122,27 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
 
   @override
   void dispose() {
-    _imageStream?.removeListener(_imageListener);
+    final listener = _imageListener;
+    if (listener != null) {
+      _imageStream?.removeListener(listener);
+    }
     _activeController.dispose();
     _settleController.dispose();
+
     super.dispose();
   }
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
+
     _getImage();
   }
 
   @override
   void didUpdateWidget(Crop oldWidget) {
     super.didUpdateWidget(oldWidget);
+
     if (widget.image != oldWidget.image) {
       _getImage();
     } else if (widget.aspectRatio != oldWidget.aspectRatio) {
@@ -167,42 +164,46 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
 
   void _getImage({bool force = false}) {
     final oldImageStream = _imageStream;
-    _imageStream = widget.image.resolve(createLocalImageConfiguration(context));
-    if (_imageStream.key != oldImageStream?.key || force) {
-      oldImageStream?.removeListener(_imageListener);
-      _imageListener =
+    final newImageStream =
+        widget.image.resolve(createLocalImageConfiguration(context));
+    _imageStream = newImageStream;
+    if (newImageStream.key != oldImageStream?.key || force) {
+      final oldImageListener = _imageListener;
+      if (oldImageListener != null) {
+        oldImageStream?.removeListener(oldImageListener);
+      }
+      final newImageListener =
           ImageStreamListener(_updateImage, onError: widget.onImageError);
-      _imageStream.addListener(_imageListener);
+      _imageListener = newImageListener;
+      newImageStream.addListener(newImageListener);
     }
   }
 
   @override
-  Widget build(BuildContext context) {
-    return ConstrainedBox(
-      constraints: const BoxConstraints.expand(),
-      child: Listener(
-        onPointerDown: (event) => pointers++,
-        onPointerUp: (event) => pointers = 0,
-        child: GestureDetector(
-          key: _surfaceKey,
-          behavior: HitTestBehavior.opaque,
-          onScaleStart: _isEnabled ? _handleScaleStart : null,
-          onScaleUpdate: _isEnabled ? _handleScaleUpdate : null,
-          onScaleEnd: _isEnabled ? _handleScaleEnd : null,
-          child: CustomPaint(
-            painter: _CropPainter(
-              image: _image,
-              ratio: _ratio,
-              view: _view,
-              area: _area,
-              scale: _scale,
-              active: _activeController.value,
+  Widget build(BuildContext context) => ConstrainedBox(
+        constraints: const BoxConstraints.expand(),
+        child: Listener(
+          onPointerDown: (event) => pointers++,
+          onPointerUp: (event) => pointers = 0,
+          child: GestureDetector(
+            key: _surfaceKey,
+            behavior: HitTestBehavior.opaque,
+            onScaleStart: _isEnabled ? _handleScaleStart : null,
+            onScaleUpdate: _isEnabled ? _handleScaleUpdate : null,
+            onScaleEnd: _isEnabled ? _handleScaleEnd : null,
+            child: CustomPaint(
+              painter: _CropPainter(
+                image: _image,
+                ratio: _ratio,
+                view: _view,
+                area: _area,
+                scale: _scale,
+                active: _activeController.value,
+              ),
             ),
           ),
         ),
-      ),
-    );
-  }
+      );
 
   void _activate() {
     _activeController.animateTo(
@@ -213,7 +214,7 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
   }
 
   void _deactivate() {
-    if (!widget.alwaysShowGrid) {
+    if (widget.alwaysShowGrid == false) {
       _activeController.animateTo(
         0.0,
         curve: Curves.fastOutSlowIn,
@@ -222,31 +223,51 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
     }
   }
 
-  Size get _boundaries =>
-      _surfaceKey.currentContext.size -
-      const Offset(_kCropHandleSize, _kCropHandleSize);
+  Size? get _boundaries {
+    final context = _surfaceKey.currentContext;
+    if (context == null) {
+      return null;
+    }
 
-  Offset _getLocalPoint(Offset point) {
-    final RenderBox box = _surfaceKey.currentContext.findRenderObject();
+    final size = context.size;
+    if (size == null) {
+      return null;
+    }
+
+    return size - const Offset(_kCropHandleSize, _kCropHandleSize) as Size;
+  }
+
+  Offset? _getLocalPoint(Offset point) {
+    final context = _surfaceKey.currentContext;
+    if (context == null) {
+      return null;
+    }
+
+    final box = context.findRenderObject() as RenderBox;
+
     return box.globalToLocal(point);
   }
 
   void _settleAnimationChanged() {
     setState(() {
       _scale = _scaleTween.transform(_settleController.value);
-      _view = _viewTween.transform(_settleController.value);
+      final nextView = _viewTween.transform(_settleController.value);
+      if (nextView != null) {
+        _view = nextView;
+      }
     });
   }
 
   Rect _calculateDefaultArea({
-    int imageWidth,
-    int imageHeight,
-    double viewWidth,
-    double viewHeight,
+    required int? imageWidth,
+    required int? imageHeight,
+    required double viewWidth,
+    required double viewHeight,
   }) {
     if (imageWidth == null || imageHeight == null) {
       return Rect.zero;
     }
+
     double height;
     double width;
     if ((widget.aspectRatio ?? 1.0) < 1) {
@@ -272,30 +293,38 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
                 viewWidth;
       }
     }
-    if (!_maxAreaWidthMap.containsKey(widget.aspectRatio)) {
-      _maxAreaWidthMap[widget.aspectRatio] = width;
+    final aspectRatio = _maxAreaWidthMap[widget.aspectRatio];
+    if (aspectRatio != null) {
+      _maxAreaWidthMap[aspectRatio] = width;
     }
+
     return Rect.fromLTWH((1.0 - width) / 2, (1.0 - height) / 2, width, height);
   }
 
   void _updateImage(ImageInfo imageInfo, bool synchronousCall) {
-    WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+    final boundaries = _boundaries;
+    if (boundaries == null) {
+      return;
+    }
+
+    WidgetsBinding.instance?.addPostFrameCallback((timeStamp) {
+      final image = imageInfo.image;
+
       setState(() {
-        _image = imageInfo.image;
+        _image = image;
         _scale = imageInfo.scale;
         _ratio = max(
-          _boundaries.width / _image.width,
-          _boundaries.height / _image.height,
+          boundaries.width / image.width,
+          boundaries.height / image.height,
         );
 
-        final viewWidth = _boundaries.width / (_image.width * _scale * _ratio);
-        final viewHeight =
-            _boundaries.height / (_image.height * _scale * _ratio);
+        final viewWidth = boundaries.width / (image.width * _scale * _ratio);
+        final viewHeight = boundaries.height / (image.height * _scale * _ratio);
         _area = _calculateDefaultArea(
           viewWidth: viewWidth,
           viewHeight: viewHeight,
-          imageWidth: _image.width,
-          imageHeight: _image.height,
+          imageWidth: image.width,
+          imageHeight: image.height,
         );
         _view = Rect.fromLTWH(
           (viewWidth - 1.0) / 2,
@@ -306,13 +335,18 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
         widget.onImageSuccess?.call();
       });
     });
-    WidgetsBinding.instance.ensureVisualUpdate();
+
+    WidgetsBinding.instance?.ensureVisualUpdate();
   }
 
-  _CropHandleSide _hitCropHandle(Offset localPoint) {
+  _CropHandleSide _hitCropHandle(Offset? localPoint) {
     final boundaries = _boundaries;
+    if (localPoint == null || boundaries == null) {
+      return _CropHandleSide.none;
+    }
+
     final viewRect = Rect.fromLTWH(
-      _boundaries.width * _area.left,
+      boundaries.width * _area.left,
       boundaries.height * _area.top,
       boundaries.width * _area.width,
       boundaries.height * _area.height,
@@ -367,38 +401,47 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
     _startView = _view;
   }
 
-  Rect _getViewInBoundaries(double scale) {
-    return Offset(
-          max(
-            min(
-              _view.left,
-              _area.left * _view.width / scale,
-            ),
-            _area.right * _view.width / scale - 1.0,
+  Rect _getViewInBoundaries(double scale) =>
+      Offset(
+        max(
+          min(
+            _view.left,
+            _area.left * _view.width / scale,
           ),
-          max(
-            min(
-              _view.top,
-              _area.top * _view.height / scale,
-            ),
-            _area.bottom * _view.height / scale - 1.0,
+          _area.right * _view.width / scale - 1.0,
+        ),
+        max(
+          min(
+            _view.top,
+            _area.top * _view.height / scale,
           ),
-        ) &
-        _view.size;
-  }
+          _area.bottom * _view.height / scale - 1.0,
+        ),
+      ) &
+      _view.size;
 
   double get _maximumScale => widget.maximumScale;
 
-  double get _minimumScale {
-    final scaleX = _boundaries.width * _area.width / (_image.width * _ratio);
-    final scaleY = _boundaries.height * _area.height / (_image.height * _ratio);
+  double? get _minimumScale {
+    final boundaries = _boundaries;
+    final image = _image;
+    if (boundaries == null || image == null) {
+      return null;
+    }
+
+    final scaleX = boundaries.width * _area.width / (image.width * _ratio);
+    final scaleY = boundaries.height * _area.height / (image.height * _ratio);
     return min(_maximumScale, max(scaleX, scaleY));
   }
 
   void _handleScaleEnd(ScaleEndDetails details) {
     _deactivate();
+    final minimumScale = _minimumScale;
+    if (minimumScale == null) {
+      return;
+    }
 
-    final targetScale = _scale.clamp(_minimumScale, _maximumScale);
+    final targetScale = _scale.clamp(minimumScale, _maximumScale);
     _scaleTween = Tween<double>(
       begin: _scale,
       end: targetScale,
@@ -418,27 +461,34 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
     );
   }
 
-  void _updateArea(
-      {double left,
-      double top,
-      double right,
-      double bottom,
-      @required _CropHandleSide cropHandleSide}) {
+  void _updateArea({
+    required _CropHandleSide cropHandleSide,
+    double? left,
+    double? top,
+    double? right,
+    double? bottom,
+  }) {
+    final image = _image;
+    if (image == null) {
+      return;
+    }
+
     var areaLeft = _area.left + (left ?? 0.0);
     var areaBottom = _area.bottom + (bottom ?? 0.0);
     var areaTop = _area.top + (top ?? 0.0);
     var areaRight = _area.right + (right ?? 0.0);
     double width = areaRight - areaLeft;
-    double height = (_image.width * _view.width * width) /
-        (_image.height * _view.height * (widget.aspectRatio ?? 1.0));
-    if (height >= 1.0 || width >= 1.0) {
+    double height = (image.width * _view.width * width) /
+        (image.height * _view.height * (widget.aspectRatio ?? 1.0));
+    final maxAreaWidth = _maxAreaWidthMap[widget.aspectRatio];
+    if ((height >= 1.0 || width >= 1.0) && maxAreaWidth != null) {
       height = 1.0;
 
       if (cropHandleSide == _CropHandleSide.bottomLeft ||
           cropHandleSide == _CropHandleSide.topLeft) {
-        areaLeft = areaRight - _maxAreaWidthMap[widget.aspectRatio];
+        areaLeft = areaRight - maxAreaWidth;
       } else {
-        areaRight = areaLeft + _maxAreaWidthMap[widget.aspectRatio];
+        areaRight = areaLeft + maxAreaWidth;
       }
     }
 
@@ -460,7 +510,8 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
     }
 
     // adjust to aspect ratio if needed
-    if (widget.aspectRatio != null && widget.aspectRatio > 0.0) {
+    final aspectRatio = widget.aspectRatio;
+    if (aspectRatio != null && aspectRatio > 0.0) {
       if (top != null) {
         areaTop = areaBottom - height;
         if (areaTop < 0.0) {
@@ -510,25 +561,36 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
     if (_action == _CropAction.cropping) {
       _lastFocalPoint = details.focalPoint;
     } else if (_action == _CropAction.moving) {
+      final image = _image;
+      if (image == null) {
+        return;
+      }
+
       final delta = details.focalPoint - _lastFocalPoint;
       _lastFocalPoint = details.focalPoint;
 
       setState(() {
         _view = _view.translate(
-          delta.dx / (_image.width * _scale * _ratio),
-          delta.dy / (_image.height * _scale * _ratio),
+          delta.dx / (image.width * _scale * _ratio),
+          delta.dy / (image.height * _scale * _ratio),
         );
       });
     } else if (_action == _CropAction.scaling) {
+      final image = _image;
+      final boundaries = _boundaries;
+      if (image == null || boundaries == null) {
+        return;
+      }
+
       setState(() {
         _scale = _startScale * details.scale;
 
-        final dx = _boundaries.width *
+        final dx = boundaries.width *
             (1.0 - details.scale) /
-            (_image.width * _scale * _ratio);
-        final dy = _boundaries.height *
+            (image.width * _scale * _ratio);
+        final dy = boundaries.height *
             (1.0 - details.scale) /
-            (_image.height * _scale * _ratio);
+            (image.height * _scale * _ratio);
 
         _view = Rect.fromLTWH(
           _startView.left + dx / 2,
@@ -542,7 +604,7 @@ class CropState extends State<Crop> with TickerProviderStateMixin, Drag {
 }
 
 class _CropPainter extends CustomPainter {
-  final ui.Image image;
+  final ui.Image? image;
   final Rect view;
   final double ratio;
   final Rect area;
@@ -550,12 +612,12 @@ class _CropPainter extends CustomPainter {
   final double active;
 
   _CropPainter({
-    this.image,
-    this.view,
-    this.ratio,
-    this.area,
-    this.scale,
-    this.active,
+    required this.image,
+    required this.view,
+    required this.ratio,
+    required this.area,
+    required this.scale,
+    required this.active,
   });
 
   @override
@@ -582,6 +644,7 @@ class _CropPainter extends CustomPainter {
 
     final paint = Paint()..isAntiAlias = false;
 
+    final image = this.image;
     if (image != null) {
       final src = Rect.fromLTWH(
         0.0,
@@ -625,7 +688,7 @@ class _CropPainter extends CustomPainter {
             boundaries.right, boundaries.top, rect.width, boundaries.bottom),
         paint);
 
-    if (!boundaries.isEmpty) {
+    if (boundaries.isEmpty == false) {
       _drawGrid(canvas, boundaries);
       _drawHandles(canvas, boundaries);
     }
